Refactor: Object relations box
