+++ mwdb/web/src/components/ShowObject/Views/RelationsBox.js
@@ -1,0 +1,131 @@
+import React, {useState, useContext} from 'react';
+import {Link} from 'react-router-dom';
+import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
+import {TagList} from '../../Tag';
+
+import api from "@mwdb-web/commons/api";
+import { ObjectContext } from "@mwdb-web/commons/context";
+import { ObjectLink, ActionCopyToClipboard } from '@mwdb-web/commons/ui';
+import RelationsAddModal from "../Actions/RelationsAddModal";
+
+function RelationsBox(props) {
+    const context = useContext(ObjectContext);
+    const [isAttributeAddModalOpen, setAttributeAddModalOpen] = useState(false);
+    const [modalError, setModalError] = useState("");
+
+    async function addObjectRelations(relation, value) {
+        try {
+            if (relation === "child")
+                await api.addObjectRelation(context.object.id, value)
+            else if (relation === "parent")
+                await api.addObjectRelation(value, context.object.id)
+            context.updateObject();
+            setAttributeAddModalOpen(false);
+        } catch(error) {
+            if(error.response && error.response.status === 404)
+                setModalError("Object not found or incorrect SHA256 hash.")
+            else
+                setModalError(error);
+        }
+    }
+
+    const parents = (props.parents || [])
+        .map((parent, index, array) =>
+            <tr key={`parent-${parent.id}`} className="flickerable">
+                <th>parent</th>
+                <td>
+                    <span>
+                        <ObjectLink {...parent}
+                                    diffWith={parent.type === "text_blob" && (array[index + 1] || {}).id}
+                                    inline/>
+                    </span>
+                    <span className="ml-2">
+                        <ActionCopyToClipboard text={parent.id} tooltipMessage="Copy sha256 to clipboard"/>
+                    </span>
+                </td>
+                <td>
+                    <TagList tags={parent.tags}/>
+                </td>
+            </tr>
+        );
+
+    const children = (props.children || [])
+        .map((child, index, array) =>
+            <tr key={`child-${child.id}`} className="flickerable">
+                <th>child</th>
+                <td>
+                    <span>
+                        <ObjectLink {...child }
+                                    diffWith={child.type === "text_blob" && (array[index + 1] || {}).id}
+                                    inline/>
+                    </span>
+                    <span className="ml-2">
+                        <ActionCopyToClipboard text={child.id} tooltipMessage="Copy sha256 to clipboard"/>
+                    </span>
+                </td>
+                <td>
+                    <TagList tags={child.tags}/>
+                </td>
+            </tr>
+        );
+    return (
+        <div className="card card-default">
+            <div className="card-header">
+                { props.header || "Relations" }
+                <Link to="#" className="float-right"
+                      onClick={(ev) => {
+                          ev.preventDefault();
+                          setAttributeAddModalOpen(true);
+                          setModalError("");
+                      }}>
+                        <FontAwesomeIcon icon="plus" pull="left" size="1x"/>
+                        Add
+                </Link>
+            </div>
+            {
+                parents.length + children.length > 0 ?
+                <table className="table table-striped table-bordered table-hover" id="rel_table">
+                    <tbody id="rel_body">
+                    { parents }
+                    { children }
+                    </tbody>
+                </table> : <div className="card-body text-muted">No relations to display</div>
+            }
+            <RelationsAddModal isOpen={isAttributeAddModalOpen}
+                               error={modalError}
+                               onSubmit={addObjectRelations}
+                               onError={setModalError}
+                               onRequestModalClose={() => setAttributeAddModalOpen(false)}
+            />
+        </div>
+    );
+}
+
+export default function MultiRelationsBox(props) {
+    const context = useContext(ObjectContext);
+
+    let TypedRelationsBox = (props) => {
+        let filterByType = function(arr) { return arr.filter(e => e.type === props.type) }
+        let filteredElements = {
+            parents: filterByType(props.parents),
+            children: filterByType(props.children)
+        }
+        if(filteredElements.parents.length + filteredElements.children.length > 0)
+            return <RelationsBox {...filteredElements} header={props.header} id={context.object.id}/>
+        else
+            return <div/>
+    }
+
+    let parents = context.object.parents;
+    let children = context.object.children;
+    let className = props.className || '';
+    return (
+        parents && children && (parents.length + children.length > 0) ?
+            <div className={className}>
+                <TypedRelationsBox header="Related samples" type="file" {...{parents, children}} />
+                <TypedRelationsBox header="Related configs" type="static_config" {...{parents, children}}/>
+                <TypedRelationsBox header="Related blobs" type="text_blob" {...{parents, children}}/>
+            </div> :
+            <RelationsBox className={className} id={context.object.id}/>
+    )
+}
